requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
dj-database-url==1.3.0
Django==4.2.2
gunicorn==20.1.0
piexif==1.1.3
pillow==10.3.0
psycopg2-binary==2.9.6
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
whitenoise==6.4.0
